Uploading files
These files are linked to the paper "Power spectrum multipoles in a curved sky: an application to the 6-degree Field Galaxy Survey".  They provide python code and data files to measure multipole power spectra, compute the convolution with a survey window function and weight function, determine the covariance of the power spectrum multipoles in a Gaussian approximation, and determine the window function multipoles.